+++ include/igraph_conversion.h
@@ -40,11 +40,11 @@ __BEGIN_DECLS
 
 IGRAPH_EXPORT igraph_error_t igraph_get_adjacency(
    const igraph_t *graph, igraph_matrix_t *res, igraph_get_adjacency_t type,
-   const igraph_vector_t *weights
+   const igraph_vector_t *weights, igraph_loops_t loops
 );
 IGRAPH_EXPORT igraph_error_t igraph_get_adjacency_sparse(
    const igraph_t *graph, igraph_sparsemat_t *res, igraph_get_adjacency_t type,
-   const igraph_vector_t *weights
+   const igraph_vector_t *weights, igraph_loops_t loops
 );
 
 IGRAPH_EXPORT igraph_error_t igraph_get_stochastic(